csv_processing/requirements.txt
google-cloud-bigquery
PyGithub
PyJWT
requests
